code_style: move some codes from `App` to `Models.InteractiveRebaseJobCollection`
